register warpscript functions
