+++ warp10/src/main/java/io/warp10/script/aggregator/modifier/NULLS_SKIP.java
@@ -17,16 +17,17 @@
 package io.warp10.script.aggregator.modifier;
 
 import io.warp10.continuum.gts.Aggregate;
+import io.warp10.continuum.gts.GeoTimeSerie;
 import io.warp10.script.NamedWarpScriptFunction;
 import io.warp10.script.WarpScriptAggregator;
-import io.warp10.script.WarpScriptAggregatorFailIfAnyNull;
+import io.warp10.script.WarpScriptAggregatorSkipIfAnyNull;
 import io.warp10.script.WarpScriptException;
 import io.warp10.script.WarpScriptLib;
 import io.warp10.script.WarpScriptStack;
 import io.warp10.script.WarpScriptStackFunction;
 
-public class NULLS_FAILIFANY extends NamedWarpScriptFunction implements WarpScriptStackFunction {
-  public NULLS_FAILIFANY(String name) {
+public class NULLS_SKIP extends NamedWarpScriptFunction implements WarpScriptStackFunction {
+  public NULLS_SKIP(String name) {
     super(name);
   }
 
@@ -38,11 +39,11 @@ public class NULLS_FAILIFANY extends NamedWarpScriptFunction implements WarpScri
       throw new WarpScriptException(getName() + " expects an AGGREGATOR");
     }
 
-    if (!(o instanceof WarpScriptAggregatorFailIfAnyNull)) {
+    if (!(o instanceof WarpScriptAggregatorSkipIfAnyNull)) {
       throw new WarpScriptException(getName() + " can not be be applied to this AGGREGATOR");
     }
 
-    WarpScriptAggregatorFailIfAnyNull aggregator = (WarpScriptAggregatorFailIfAnyNull) o;
+    WarpScriptAggregatorSkipIfAnyNull aggregator = (WarpScriptAggregatorSkipIfAnyNull) o;
 
     if (aggregator.actionOnNullsIsSet()) {
       throw new WarpScriptException(getName() + " can not be applied on an AGGREGATOR that already has an action on null values");
@@ -53,18 +54,20 @@ public class NULLS_FAILIFANY extends NamedWarpScriptFunction implements WarpScri
     return stack;
   }
 
-  private static final class ModifiedAggregator extends NamedWarpScriptFunction implements WarpScriptAggregatorFailIfAnyNull {
+  private static final class ModifiedAggregator extends NamedWarpScriptFunction implements WarpScriptAggregatorSkipIfAnyNull {
 
-    private final WarpScriptAggregatorFailIfAnyNull aggregator;
+    private final WarpScriptAggregatorSkipIfAnyNull aggregator;
 
-    public ModifiedAggregator(String name, WarpScriptAggregatorFailIfAnyNull aggregator) {
+    public ModifiedAggregator(String name, WarpScriptAggregatorSkipIfAnyNull aggregator) {
       super(name);
       this.aggregator = aggregator;
     }
 
     @Override
     public Object apply(Aggregate aggregate) throws WarpScriptException {
-      WarpScriptAggregatorFailIfAnyNull.failIfAnyNull(aggregate);
+      if (WarpScriptAggregatorSkipIfAnyNull.containsAnyNull(aggregate)) {
+        return new Object[] { Long.MAX_VALUE, GeoTimeSerie.NO_LOCATION, GeoTimeSerie.NO_ELEVATION, null };
+      }
       return aggregator.apply(aggregate);
     }
